+++ sl_controller/terminal.lua
@@ -24,8 +24,9 @@ Local commands:
 - help     = this message 
 - pub      = switch to public use
 - priv      = switch to private use
-Test commands:
-- send <num> on/off   = send on/off event
+Global commands:
+- send <num> on/off  = send on/off event
+- msg <num> <text>    = send a text message
 ]]
 
 local function formspec1()
@@ -82,16 +83,25 @@ local function command(pos, cmnd, player)
 			output(pos, player..":$ "..cmnd)
 			output(pos, "Switched to private use!")
 		elseif meta:get_int("public") == 1 or owner == player then
+			-- send <num> on/off
 			local num, topic = cmnd:match('^send%s+([0-9]+)%s+([onff]+)$')
 			if num and topic then
 				local own_number = meta:get_string("own_number")
 				output(pos, player..":$ send "..num.." "..topic)
 				tubelib.send_message(num, owner, nil, topic, own_number)
-			else
-				local number = meta:get_string("number") or "0000"
-				output(pos, player..":$ "..cmnd)
-				tubelib.send_message(number, owner, nil, "term", cmnd)
+				return
 			end
+			-- msg <num> <text>
+			local num, text = cmnd:match('^msg%s+([0-9]+)%s+(.+)$')
+			if num and text then
+				local own_number = meta:get_string("own_number")
+				output(pos, player..":$ msg "..num.." "..text)
+				tubelib.send_message(num, owner, nil, "msg", {src=own_number, text=text})
+				return
+			end
+			local number = meta:get_string("number") or "0000"
+			output(pos, player..":$ "..cmnd)
+			tubelib.send_message(number, owner, nil, "term", cmnd)
 		end
 	end
 end	
@@ -203,11 +213,70 @@ minetest.register_craft({
 	},
 })
 
-tubelib.register_node("sl_controller:terminal", {"sl_controller:terminal2", "sl_controller:terminal3"}, {
+tubelib.register_node("sl_controller:terminal", {}, {
 	on_recv_message = function(pos, topic, payload)
 		if topic == "term" then
 			output(pos, payload)
 			return true
+		elseif topic == "msg" then
+			output(pos, payload.src..": "..payload.text)
+			return true
+		end
+	end,
+})
+
+tubelib.register_node("sl_controller:terminal2", {}, {
+	on_recv_message = function(pos, topic, payload)
+		if topic == "term" then
+			output(pos, payload)
+			return true
+		elseif topic == "msg" then
+			output(pos, payload.src..": "..payload.text)
+			return true
 		end
 	end,
 })		
+
+sl_controller.register_function("get_term", {
+	cmnd = function(self)
+		return sl_controller.get_command(self.meta.number)
+	end,
+	help = ' $get_term()  --> text string or nil\n'..
+		' Read an entered string (command) from the Terminal.\n'..
+		' example: s = $get_term()\n'..
+		" The Terminal has to be connected to the controller."
+})
+
+sl_controller.register_action("put_term", {
+	cmnd = function(self, num, text)
+		text = tostring(text or "")
+		tubelib.send_message(num, self.meta.owner, nil, "term", text)
+	end,
+	help = " $put_term(num, text)\n"..
+		' Send a text line to the terminal with number "num".\n'..
+		' example: $put_term("0123", "Hello "..name)'
+})
+
+sl_controller.register_function("get_msg", {
+	cmnd = function(self)
+		local msg = sl_controller.get_msg(self.meta.number)
+		if msg then
+			return msg.src, msg.text
+		end
+	end,
+	help = ' $get_msg()  --> number and text string or nil\n'..
+		' Read a received messages. Number is the node\n'..
+		' number of the sender.\n'..
+		' example: num,msg = $get_msg().'
+})
+
+sl_controller.register_action("send_msg", {
+	cmnd = function(self, num, text)
+		local msg = {src = self.meta.number, text = tostring(text or "")}
+		tubelib.send_message(num, self.meta.owner, nil, "msg", msg)
+	end,
+	help = " $send_msg(num, text)\n"..
+		' Send a message to the controller with number "num".\n'..
+		' example: $send_msg("0123", "test")'
+})
+